feat: add start option to post build (#151)
