+++ package.json
@@ -6,6 +6,7 @@
   "main": "src",
   "scripts": {
     "test": "cypress run",
+    "test:unit": "jest",
     "semantic-release": "semantic-release",
     "start": "serve public"
   },
@@ -41,6 +42,7 @@
     "node": ">=10.18.1"
   },
   "devDependencies": {
+    "jest": "^24.9.0",
     "cypress": "6.8.0",
     "netlify-cli": "3.13.7",
     "prettier": "2.2.1",
@@ -61,5 +63,8 @@
         "prerelease": true
       }
     ]
+  },
+  "jest": {
+    "rootDir": "./src"
   }
 }
