+++ src/constants.js
@@ -1,0 +1,4 @@
+module.exports = {
+  PLUGIN_NAME: 'netlify-plugin-cypress',
+  DEFAULT_BROWSER: 'chromium',
+}
