+++ tests/test-postbuild-start/README.md
@@ -1,0 +1,10 @@
+# run post-build tests only with start command
+
+These tests run:
+- [ ] before the build
+- [x] after the build
+- [ ] on deploy success
+
+## Local build
+
+Use command `DEBUG=netlify-plugin-cypress npm run netlify:build` to try the plugin locally
