+++ tests/test-postbuild-start/cypress.json
@@ -1,0 +1,7 @@
+{
+  "pluginsFile": false,
+  "supportFile": false,
+  "fixturesFolder": false,
+  "baseUrl": "http://localhost:5000",
+  "projectId": "ixroqc"
+}
