+++ tests/test-postbuild-start/cypress/integration/spec.js
@@ -1,0 +1,5 @@
+/// <reference types="cypress" />
+it('loads page', () => {
+  cy.visit('/')
+  cy.contains('Placeholder').should('be.visible')
+})
