+++ tests/test-postbuild-start/netlify.toml
@@ -1,0 +1,32 @@
+[build]
+command = "echo 'Netlify build command ...'"
+publish = "public"
+
+[build.environment]
+# cache Cypress binary in local "node_modules" folder
+# so Netlify caches it
+CYPRESS_CACHE_FOLDER = "./node_modules/CypressBinary"
+# set TERM variable for terminal output
+TERM = "xterm"
+# do not print too many progress messages
+CI = "1"
+
+[[plugins]]
+  # local Cypress plugin will test our site after it is built
+  # in production, please use: package = "netlify-plugin-cypress"
+  package = "../../"
+
+  # run Cypress tests once on the site before it is built
+  # and do not run the tests after it was built
+
+  [plugins.inputs]
+    # these settings apply to the default step onSuccess
+    # in this case, we do not want to run any tests after deploy
+    enable = false
+
+  # let's run tests against development server
+  [plugins.inputs.postBuild]
+    enable = true
+    start = 'npx serve public'
+    wait-on = 'http://localhost:5000'
+    wait-on-timeout = '30' # seconds
