+++ tests/test-postbuild-start/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "postbuild-start-example",
+  "version": "0.1.0",
+  "private": true,
+  "scripts": {
+    "netlify:build": "../../node_modules/.bin/netlify build"
+  }
+}
