+++ tests/test-postbuild-start/public/index.html
@@ -1,0 +1,1 @@
+<body>Placeholder</body>
